backend/requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
python-dotenv==0.19.1
bcrypt==3.2.0
mysqlclient==2.0.3
PyMySQL==1.0.2
pymongo==3.12.1
eventlet==0.30.2
dnspython==1.16.0
Werkzeug==2.0.3
SQLAlchemy==1.4.46
pysuricata==0.0.4
scapy==2.4.5
pycryptodomex==3.11.0
redis==4.0.2
python-magic==0.4.24
pika==1.2.0
